Use `cython.floating` instead of `anyfloat`
Cython has a built-in fused type `floating` representing both `float` and `double`. It's the same as the custom-defined `anyfloat`. I think it's better to use the built-in so we don't need to define `anyfloat` here and there.

See http://cython.readthedocs.io/en/latest/src/userguide/fusedtypes.html#built-in-fused-types for reference.